Refactor/Build: Add separate signup forms for students and professors
Changed signup.html into a form to select type of user then redirecting
to a separate page for each user type. This allows for the submission
of the specific fields such as the codice for the student. Added a php
file which allows for the logic associated and built the signup html
pages.